fix: fix dismissible banner ssr bug (#547)
