+++ packages/data-explorer-ui/src/hooks/useFeatureFlag/common/utils.ts
@@ -1,3 +1,5 @@
+import { setLocalStorage } from "../../useLocalStorage/common/utils";
+
 /**
  * Set feature flags from URL.
  * @param features - List of feature flags.
@@ -12,24 +14,4 @@ export function setFeatureFlags(features: string[]): void {
       setLocalStorage(key, value);
     }
   }
-}
-
-/**
- * Return the value for the specified key.
- * @param key - Key.
- * @returns value.
- */
-export function getLocalStorage(key: string): string | null {
-  if (typeof window === "undefined") return null;
-  return window?.localStorage?.getItem(key) ?? null;
-}
-
-/**
- * Set the value for the specified key.
- * @param key - Key.
- * @param value - Value.
- */
-export function setLocalStorage(key: string, value: string): void {
-  if (typeof window === "undefined") return;
-  window?.localStorage?.setItem(key, value);
 }
