+++ packages/data-explorer-ui/src/hooks/useFeatureFlag/useFeatureFlag.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
+import { getLocalStorage } from "../useLocalStorage/common/utils";
 import { FLAG } from "./common/entities";
-import { getLocalStorage } from "./common/utils";
 
 /**
  * Determine if feature is available to user.
@@ -8,15 +8,12 @@ import { getLocalStorage } from "./common/utils";
  * @returns true if feature is available to user.
  */
 export function useFeatureFlag(featureFlag: string): boolean {
-  /* Flag value from local storage. */
-  const enabled = getLocalStorage(featureFlag) === FLAG.TRUE;
-  /* Flag indicating if feature is available to user. */
-  const [isEnabled, setIsEnabled] = useState<boolean>(enabled);
+  const [isEnabled, setIsEnabled] = useState<boolean>(false);
 
-  /* Update state of enabled flag and redirect user if feature is not available to them. */
+  // Sets state of enabled flag.
   useEffect(() => {
-    setIsEnabled(enabled);
-  }, [enabled, featureFlag]);
+    setIsEnabled(getLocalStorage(featureFlag) === FLAG.TRUE);
+  }, [featureFlag]);
 
   return isEnabled;
 }
